// AI_Workspace/pixel-agents-zcorvus/webview-ui/src/office/types.ts
export const CharacterState = {
  IDLE: 'idle',
  WALK: 'walk',
  TYPE: 'type',
} as const;
export type CharacterState = (typeof CharacterState)[keyof typeof CharacterState];